Delete unused files and reorganize repository files
