+++ app/src/androidTest/kotlin/com/jehutyno/yomikata/repository/local/RoomTestSampleEntities.kt
@@ -1,5 +1,8 @@
 package com.jehutyno.yomikata.repository.local
 
+import com.jehutyno.yomikata.dao.QuizDao
+import com.jehutyno.yomikata.dao.WordDao
+
 
 /**
  * Samples of Room entities for testing
