+++ app/src/androidTest/kotlin/com/jehutyno/yomikata/repository/local/StatsDaoTest.kt
@@ -4,6 +4,7 @@ import androidx.room.Room
 import androidx.test.core.app.ApplicationProvider
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.filters.SmallTest
+import com.jehutyno.yomikata.dao.StatsDao
 import org.junit.Assert.*
 
 import org.junit.After
