+++ app/src/main/kotlin/com/jehutyno/yomikata/dao/KanjiSoloDao.kt
@@ -1,8 +1,11 @@
-package com.jehutyno.yomikata.repository.local
+package com.jehutyno.yomikata.dao
 
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.Query
+import com.jehutyno.yomikata.repository.local.RoomKanjiSolo
+import com.jehutyno.yomikata.repository.local.RoomKanjiSoloRadical
+import com.jehutyno.yomikata.repository.local.RoomRadicals
 
 
 @Dao
