+++ app/src/main/kotlin/com/jehutyno/yomikata/dao/QuizDao.kt
@@ -1,6 +1,8 @@
-package com.jehutyno.yomikata.repository.local
+package com.jehutyno.yomikata.dao
 
 import androidx.room.*
+import com.jehutyno.yomikata.repository.local.RoomQuiz
+import com.jehutyno.yomikata.repository.local.RoomQuizWord
 
 
 @Dao
