+++ app/src/main/kotlin/com/jehutyno/yomikata/dao/SentenceDao.kt
@@ -1,9 +1,10 @@
-package com.jehutyno.yomikata.repository.local
+package com.jehutyno.yomikata.dao
 
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.Query
 import androidx.room.Update
+import com.jehutyno.yomikata.repository.local.RoomSentences
 
 
 @Dao
