+++ app/src/main/kotlin/com/jehutyno/yomikata/repository/local/SentenceSource.kt
@@ -1,6 +1,7 @@
 package com.jehutyno.yomikata.repository.local
 
 import android.database.sqlite.SQLiteDatabase
+import com.jehutyno.yomikata.dao.SentenceDao
 import com.jehutyno.yomikata.model.Sentence
 import com.jehutyno.yomikata.model.Word
 import com.jehutyno.yomikata.repository.SentenceRepository
